Se corrigio el precio

+++ src/main/java/org/tecsharp/apiservlet/webapp/headers/models/ItemCarro.java
@@ -36,7 +36,7 @@ public class ItemCarro {
                 && Objects.equals(producto.getNombre(), itemCarro.producto.getNombre());
     }
 
-    public Double getImporte() {
+    public Integer getImporte() {
         return cantidad * producto.getPrecio();
     }
 }
